rename tests project.
