+++ CredentialsManagerTests/CredentialsManagerTests.csproj
