+++ CredentialsManagerTests/Files/Input/read1-CRLF.txt
