+++ CredentialsManagerTests/Files/Input/read2-LF.txt
