+++ CredentialsManagerTests/Files/Output/expected_multi_line.txt
